requirements_test.txt
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
django>=2.2
djangorestframework>=3.12
django-appconf>=1.0.4
Pillow>=8.1.0
